Rebuild `pip`
Appears that some old `conda-build` issues caused `pip` to include the
`pyc` files from `setuptools`. As this problem has long since been fixed
in `conda-build`, this merely triggers a rebuild of `pip` to create a
new package that does not have these unrelated files. The affected
packages seem to only be on Python 3.5 as the Python 3.6 packages were
built after the `conda-build` fix and the Python 2.7 packages used an
older structure for `.pyc` files, which `conda-build` already handled
correctly.